pyproject.toml
[tool.poetry]
name = "suspiciousduck"
version = "0.1.0"
description = "A suspicious bot. To be honest, it's not that suspicious."
authors = ["illumfx <illumfx@mail.de>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
python-dotenv = "^0.18.0"
pendulum = "^2.1.2"
parsedatetime = "^2.6"
asyncpg = "^0.23.0"
twitchio = "^1.2.3"
libneko = {git = "https://gitlab.com/Tmpod/libneko"}

[tool.poetry.dev-dependencies]
black = {version = "^21.6b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
